+++ packages/terra-icon/src/icon/IconDiamondLowLight.jsx
@@ -1,20 +1,26 @@
 /* eslint-disable */
 import React from 'react';
-import classNames from 'classnames/bind';
+import classNames from 'classnames';
+import classNamesBind from 'classnames/bind';
+import ThemeContext from 'terra-theme-context';
 import IconBase from '../IconBase';
 import styles from '../IconDiamondLowLight.module.scss';
 
-const cx = classNames.bind(styles);
+const cx = classNamesBind.bind(styles);
 
 const SvgIcon = (customProps) => {
   const attributes = Object.assign({}, customProps);
-  const iconClassNames = cx([
-    'IconDiamondLowLight',
+  const theme = React.useContext(ThemeContext);
+  const iconClassNames = classNames(
+    cx(
+      'IconDiamondLowLight',
+      theme.className,
+    ),
     customProps.className,
-  ]);
+  );
 
   return (
-    <IconBase className={iconClassNames} {...attributes}>
+    <IconBase {...attributes} className={iconClassNames} >
       <path fill="#C270D7" d="M24.8 47.7l22.9-22.9c.4-.4.4-1.1 0-1.6L24.8.3c-.4-.4-1.1-.4-1.6 0L.3 23.2c-.4.4-.4 1.1 0 1.6l22.9 22.9c.4.4 1.2.4 1.6 0z" ></path>
     </IconBase>
   );