+++ packages/terra-icon/CHANGELOG.md
@@ -6,6 +6,10 @@ Changed
 
 * Update changelog formatting
 
+Fixed
+
+* Fixed how class names are identified with css modules to support themes properly
+
 ## 3.33.0 - (July 7, 2020)
 
 Changed